requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
python-dotenv
PyMySQL
pandas
plotly
reportlab
cryptography
openpyxl
Flask-Migrate